feat(sync): Add checks when `containerId` is provided
I run into an issue providing a container ID of an detached container.
With the implementation before I got an `Uncaught TypeError` since there
is no type check whether `getContainerId()` returns a container so that
`map.get()` cannot be called. I took me a while to find the issue.

Thus I suggest to add some type checks when `containerId` is provided
with appropriate error message. This should result in a better developer
experience.